+++ backend-spring/src/main/java/com/example/factoryguard/application/service/inspection/SubmitRealtimeInspectionService.java
@@ -36,6 +36,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 import java.util.concurrent.TimeoutException;
@@ -89,7 +90,7 @@ public class SubmitRealtimeInspectionService implements SubmitRealtimeInspection
                 .sourceType("CAMERA")
                 .sourceId(String.valueOf(camera.getCameraId()))
                 .runStatus(RunStatus.PENDING)
-                .appliedThreshold(resolved.getAnomalyThreshold())
+                .appliedThreshold(BigDecimal.valueOf(resolved.getAnomalyThreshold()))
                 .idempotencyKey(UUID.randomUUID().toString())
                 .startedAt(LocalDateTime.now())
                 .build());
